requirements.txt
streamlit
google-generativeai
python-dotenv>=1.0
requests>=2.31
